protobuf,examples: move json encoding to examples
